[INJIWEB-1429] use redux persist library to persist the issuers list
Currently we are storing the selected issuer details in redux store and using it when user selected a specific issuer to download the VC. When user is navigated to auth server web page and clicks on back button, instead of showing selected issuer credential types it is showing no credentials found. To resolve this we are using redux persist to persist the issuers list.

Signed-off-by: PuBHARGAVI <46226958+PuBHARGAVI@users.noreply.github.com>